+++ packages/terra-functional-testing/src/commands/utils/cleanScreenshots.js
@@ -10,9 +10,8 @@ const isDirectory = filePath => (fs.existsSync(filePath) && fs.lstatSync(filePat
 
 /**
  * Delete the `diff`, `error`, `latest` and `reference` screenshot directories for each test run so all the screenshots are current.
- * @param {boolean} cleanReferenceScreenshots - A flag to determine if the reference screenshots should also be deleted since they will be downloaded from the remote repository.
  */
-async function cleanScreenshots(cleanReferenceScreenshots) {
+async function cleanScreenshots() {
   const monoRepoPath = path.resolve(process.cwd(), 'packages');
   const isMonoRepo = fs.existsSync(monoRepoPath);
   const patterns = [];
@@ -25,19 +24,11 @@ async function cleanScreenshots(cleanReferenceScreenshots) {
       patterns.push(path.resolve(monoRepoPath, packageName, 'tests', 'wdio', '**', '__snapshots__', 'diff'));
       patterns.push(path.resolve(monoRepoPath, packageName, 'tests', 'wdio', '**', '__snapshots__', 'error'));
       patterns.push(path.resolve(monoRepoPath, packageName, 'tests', 'wdio', '**', '__snapshots__', 'latest'));
-
-      if (cleanReferenceScreenshots) {
-        patterns.push(path.resolve(monoRepoPath, packageName, 'tests', 'wdio', '**', '__snapshots__', 'reference'));
-      }
     });
   } else {
     patterns.push(path.resolve(process.cwd(), 'tests', 'wdio', '**', '__snapshots__', 'diff'));
     patterns.push(path.resolve(process.cwd(), 'tests', 'wdio', '**', '__snapshots__', 'error'));
     patterns.push(path.resolve(process.cwd(), 'tests', 'wdio', '**', '__snapshots__', 'latest'));
-
-    if (cleanReferenceScreenshots) {
-      patterns.push(path.resolve(process.cwd(), 'tests', 'wdio', '**', '__snapshots__', 'reference'));
-    }
   }
 
   // For each of the three directory patterns, do a glob search to identify all the existing and matching screenshot directories and delete them.