+++ packages/terra-functional-testing/src/services/wdio-terra-service.js
@@ -1,12 +1,15 @@
 /* eslint-disable class-methods-use-this */
 const expect = require('expect');
+const { SevereServiceError } = require('webdriverio');
 const { accessibility, element, screenshot } = require('../commands/validates');
 const { toBeAccessible, toMatchReference } = require('../commands/expect');
+const getRemoteScreenshotConfiguration = require('../config/utils/getRemoteScreenshotConfiguration');
 const {
   describeTests,
   describeViewports,
   getViewports,
   hideInputCaret,
+  ScreenshotRequestor,
   setApplicationLocale,
   setViewport,
 } = require('../commands/utils');
@@ -26,6 +29,23 @@ class TerraService {
       ...launcherOptions,
       ...serviceOptions,
     };
+  }
+
+  /**
+   * Gets executed once before all workers get launched.
+   * Downloads the reference screenshots from the remote repository if useRemoteReferenceScreenshots is true.
+   * @param {Object} config wdio configuration object
+   */
+  async onPrepare(config) {
+    try {
+      if (this.serviceOptions.useRemoteReferenceScreenshots) {
+        const screenshotConfig = getRemoteScreenshotConfiguration(config.screenshotsSites, this.serviceOptions.buildBranch);
+        const screenshotRequestor = new ScreenshotRequestor(screenshotConfig.publishScreenshotConfiguration);
+        await screenshotRequestor.download();
+      }
+    } catch (error) {
+      throw new SevereServiceError(error);
+    }
   }
 
   /**